+++ notes.md
@@ -10,5 +10,11 @@ ideas/TODO:
 
 [ ] 01-2020 change the Orangish color, replace with something more muted (i like the purple lol)
 
+[] add a sticky nav, as scroll down floats to the top and then title name shrinks to the right
+
+[] slow transition on load
+
+[] remove design detail make more simple
+
 start building post
 sanity start
