+++ web/pages/about.js
@@ -2,7 +2,23 @@ import Layout from "../component/Layout";
 const PageName = () => {
   return (
     <div>
-      <h2 className="subtitle">A developer and other</h2>
+      <div className="description">
+        <h2>Developer and problem solver</h2>
+        <p>
+          Matthew Atterholt, <br /> I enjoy using programing to solve problems.
+          I find myself using Python to create small application. I being a user
+          and a developer, communicate with the team to insure that my
+          application will to assist the them to complete task and ship results.
+          I know a bit about front-end development and building web apps with
+          React.js. I have dabble back-end development using Flask and
+          Express.js to build REST API. To keep all the data organize I worked
+          with building relational tables using sqlite and PostgreSQL.
+        </p>
+        <p>
+          Check out my Projects to see what I have built or my Notes to see the
+          things that I have learn
+        </p>
+      </div>
     </div>
   );
 };
@@ -11,6 +27,7 @@ const Content = () => {
     <>
       <PageName />
       <div className="description">
+        <h2>Other then developer</h2>
         <p>
           Work as a Validation Technician for 14 years. Started to learn Python
           to help automate task that would slow down and annoy the team. As my
@@ -31,7 +48,7 @@ const Content = () => {
 function AboutPage() {
   return (
     <main className="main">
-      <Layout children={<Content />} subtitle="About" />
+      <Layout children={<Content />} />
     </main>
   );
 }
