+++ web/style/header.module.css
@@ -1,0 +1,92 @@
+.header_container {
+    overflow: hidden;
+    height: 20%;
+    width: 100%;
+}
+
+.header {
+    display: grid;
+    grid-template-columns: repeat(8, 100px);
+    grid-template-rows: repeat(5, 100px);
+    letter-spacing: 2px;
+    font-family: 'Bangers',
+        cursive;
+    margin: 0 25px;
+    font-size: 50px;
+    font-weight: bolder;
+    transform: rotate(15deg);
+}
+
+.headOne {
+    grid-area: head1;
+    align-self: center;
+    justify-self: center;
+    transform: rotate(-15deg);
+    z-index: 10;
+    grid-column: 3;
+    grid-row: 2;
+}
+
+.headTwo {
+    grid-area: head2;
+    align-self: center;
+    justify-self: center;
+    transform: rotate(-15deg);
+    z-index: 10;
+    grid-column: 5;
+    grid-row: 3;
+}
+
+.bg_pattern {
+    background-image: url("data:image/svg+xml,%3Csvg width='6' height='6' viewBox='0 0 6 6' xmlns='http://www.w3.org/2000/svg'%3E%3Cg fill='%23ffc857' fill-opacity='0.4' fill-rule='evenodd'%3E%3Cpath d='M5 0h1L0 6V5zM6 5v1H5z'/%3E%3C/g%3E%3C/svg%3E");
+    grid-column: 6;
+    grid-row: 1;
+}
+
+.bg_pattern:nth-child(3) {
+    grid-column: 2;
+    grid-row: 2;
+
+}
+
+.bg_pattern:nth-child(4) {
+    grid-column: 6;
+    grid-row: 1;
+}
+
+.bg_pattern:nth-child(5) {
+    grid-column: 3;
+    grid-row: 4;
+}
+
+.bg_pattern:nth-child(6) {
+    grid-column: 5;
+    grid-row: 4;
+}
+
+.bg_orange {
+    background-color: var(--main_orange);
+    grid-column: 8;
+    grid-row: 1;
+}
+
+.bg_orange:nth-child(7) {
+
+
+    grid-column: 7;
+    grid-row: 3;
+
+
+}
+
+.bg_orange:nth-child(8) {
+
+    grid-column: 1;
+    grid-row: 4;
+
+}
+
+.bg_orange:nth-child(9) {
+    grid-column: 4;
+    grid-row: 1;
+}
