+++ web/style/index.module.css
@@ -6,7 +6,6 @@
     width: 100%;
 }
 
-
 .name {
     color: var(--main-blue);
     margin: 5px 10px;
